SNOWYAPP/requirements.txt
aiogram>=3.4.0
asyncpg>=0.29.0
python-dotenv>=1.0.1
fastapi>=0.110.0
uvicorn>=0.27.0